pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "getit"
version = "0.1.0"
description = "Universal file hosting downloader with TUI - supports GoFile, PixelDrain, MediaFire, 1Fichier, Mega.nz"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "getit contributors" }
]
keywords = [
    "download",
    "file-hosting",
    "gofile",
    "pixeldrain",
    "mediafire",
    "mega",
    "1fichier",
    "cli",
    "tui",
    "async",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Utilities",
]
dependencies = [
    "aiohttp>=3.9.0",
    "aiofiles>=24.1.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "textual>=0.70.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.0",
    "aiosqlite>=0.20.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "aiolimiter>=1.1.0",
    "tenacity>=8.3.0",
    "pycryptodomex>=3.20.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
getit = "getit.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/getit"
Repository = "https://github.com/yourusername/getit"
Issues = "https://github.com/yourusername/getit/issues"

[tool.hatch.build.targets.sdist]
include = [
    "/src/getit",
]

[tool.hatch.build.targets.wheel]
packages = ["src/getit"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest-asyncio>=1.3.0",
]
